hide pods/containers instead of removing from model
It seems that the API server 'remembers' deleted pods for quite a long
time, we have seen an event for a deleted pod 5h ago that was still
received and rendered in podview.

We may experience churn in the model, where our garbage collection
removes old pods but the event stream potentially adds them back in. If
we have logic in the future that depends on how long we've known that a
pod/container has been in a certain state, based on our local state,
this would not work.

So it seems better just to hide old pods/containers instead, which is
what we do here.